[ fix ] bring back clause tracing
