+++ Src/Term/Display.hs
@@ -5,6 +5,7 @@ module Term.Display where
 import Display
 import Term
 import Thin
+import Unelaboration.Monad
 import Unelaboration
 
 instance (Show m, UnelabMeta m) => Display (Tm m) where
